Added additional check on Phleg
Added PrePhlegSplit Variable. This variable is required to be true for splitting Phleg. This fixes the problem of dying and then the autosplit splitting on Phleg. The variable will be set to true whenever the player reaches the phone screen with Phleg, so the autosplit won't split if the player dies at the last tier of slogs.